requirements.txt
Django==2.2
djangorestframework==3.9.4
Pillow==6.0.0
python-dotenv==0.10.3
django-filter==2.2.0
celery==4.3.0
redis==3.3.11
djangorestframework-simplejwt==3.3
pytest==5.2.2
pytest-django==3.6.0
factory_boy==2.12.0
Faker==2.0.3
